Use existing span when adding xray context to events
