+++ internal/event/client.go
@@ -9,8 +9,6 @@ import (
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/eventbridge"
 	"github.com/aws/aws-sdk-go-v2/service/eventbridge/types"
-	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/trace"
 )
 
 const source = "opg.poas.makeregister"
@@ -89,12 +87,6 @@ func send[T any](ctx context.Context, c *Client, detail any) error {
 	if !ok {
 		return errors.New("event send of unknown type")
 	}
-
-	tracer := otel.GetTracerProvider().Tracer("mlpab")
-	ctx, span := tracer.Start(ctx, detailType,
-		trace.WithSpanKind(trace.SpanKindInternal),
-	)
-	defer span.End()
 
 	v, err := json.Marshal(detail)
 	if err != nil {
