+++ internal/telemetry/telemetry.go
@@ -5,12 +5,10 @@ package telemetry
 import (
 	"context"
 	"fmt"
-	"log/slog"
 	"net/http"
 	"strings"
 
 	"github.com/felixge/httpsnoop"
-	"github.com/ministryofjustice/opg-modernising-lpa/internal/appcontext"
 	"go.opentelemetry.io/contrib/propagators/aws/xray"
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
@@ -76,43 +74,4 @@ func WrapHandler(handler http.Handler) http.HandlerFunc {
 		span.SetStatus(semconv.SpanStatusFromHTTPStatusCodeAndSpanKind(m.Code, trace.SpanKindServer))
 		span.SetAttributes(semconv.HTTPResponseContentLengthKey.Int64(m.Written))
 	}
-}
-
-type SlogHandler struct {
-	handler slog.Handler
-}
-
-func NewSlogHandler(h slog.Handler) slog.Handler {
-	return &SlogHandler{handler: h}
-}
-
-func (h *SlogHandler) Enabled(ctx context.Context, level slog.Level) bool {
-	return h.handler.Enabled(ctx, level)
-}
-
-func (h *SlogHandler) Handle(ctx context.Context, record slog.Record) error {
-	spanCtx := trace.SpanContextFromContext(ctx)
-	if spanCtx.HasTraceID() {
-		traceID := spanCtx.TraceID()
-		record.AddAttrs(slog.String("trace_id", traceID.String()))
-	}
-
-	session, err := appcontext.SessionFromContext(ctx)
-	if err == nil {
-		record.AddAttrs(slog.String("session_id", session.SessionID))
-
-		if session.OrganisationID != "" {
-			record.AddAttrs(slog.String("organisation_id", session.OrganisationID))
-		}
-	}
-
-	return h.handler.Handle(ctx, record)
-}
-
-func (h *SlogHandler) WithAttrs(attrs []slog.Attr) slog.Handler {
-	return NewSlogHandler(h.handler.WithAttrs(attrs))
-}
-
-func (h *SlogHandler) WithGroup(name string) slog.Handler {
-	return NewSlogHandler(h.handler.WithGroup(name))
 }
